# docs/swagger-spec.yaml
{"openapi":"3.0.0","paths":{"/b2b/hello/public":{"get":{"operationId":"AppController_getHelloPublic","parameters":[],"responses":{"200":{"description":"The app it's alive!"}},"tags":["** hello **"],"security":[{"bearer":[]}]}},"/b2b/hello/private":{"get":{"operationId":"AppController_getHelloPrivate","parameters":[],"responses":{"200":{"description":"The app with verification it's alive!"}},"tags":["** hello **"],"security":[{"bearer":[]}]}},"/b2b/claims":{"get":{"operationId":"ClaimsController_get","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"id","required":false,"in":"query","schema":{"type":"number"}},{"name":"businessUnit","required":false,"in":"query","schema":{"type":"array","items":{"type":"number"}}},{"name":"userInCharge","required":false,"in":"query","schema":{"type":"array","items":{"type":"number"}}},{"name":"status","required":false,"in":"query","schema":{"type":"array","items":{"type":"number"}}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseMessageDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomResponse"}}}}},"tags":["Claims"],"security":[{"bearer":[]}]}}},"info":{"title":"poc-bff-nest","description":"Bff mega description","version":"v1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"Authorization":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"ErrorObject":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"CustomResponse":{"type":"object","properties":{"timestamp":{"type":"string"},"error":{"$ref":"#/components/schemas/ErrorObject"}},"required":["timestamp","error"]},"LabelDto":{"type":"object","properties":{"en_name":{"type":"string","description":"english name"},"es_name":{"type":"string","description":"spanish name"}},"required":["en_name","es_name"]},"ResponsePaginatorDto":{"type":"object","properties":{"currentPage":{"type":"number"},"totalPages":{"type":"number"},"nextPage":{"type":"number"},"previousPage":{"type":"number"},"totalRecords":{"type":"number"}},"required":["currentPage","totalPages","nextPage","previousPage","totalRecords"]},"ResponseMessageDto":{"type":"object","properties":{"statusCode":{"type":"number"},"data":{"type":"object"},"message":{"$ref":"#/components/schemas/LabelDto"},"paginator":{"$ref":"#/components/schemas/ResponsePaginatorDto"}},"required":["statusCode","data","message","paginator"]}}}}